refactor: move asking for camera permissions into IdScanner
